Fix source-size gate: trim parse_pr.rs and messages.rs under 1000 lines
parse_pr.rs (1018) and messages.rs (1005) exceeded the 1000-line hard
limit after #486/#487 landed on main, breaking the source-size gate on
every PR rebased onto current main (including #498).

- parse_pr.rs: extract mergeable_parse_tests into parse_pr_tests.rs,
  matching the existing foo.rs/foo_tests.rs convention used by
  edit_properties and pr_diff.
- messages.rs: condense verbose module/variant doc comments (no logic
  change).
